+++ l10n/ru.js
@@ -119,7 +119,6 @@ OC.L10N.register(
     "Anonymous": "Аноним",
     "Default editor theme": "Тема редактора по умолчанию",
     "Light": "Светлая",
-    "Classic Light": "Классическая светлая",
     "Dark": "Темная",
     "Review only": "Только рецензирование",
     "Comment only": "Только комментирование",
@@ -155,8 +154,6 @@ OC.L10N.register(
     "Select file to insert text" : "Выбрать файл для вставки текста",
     "Select file" : "Выбрать файл",
     "Enable sharing (might increase editors loading time)" : "Включить панель общего доступа (может увеличить время загрузки редакторов)",
-    "Same as system" : "Такая же, как в системе",
-    "Contrast dark" : "Контрастная темная",
-    "Gray" : "Серая"
+    "Same as system" : "Такая же, как в системе"
 },
 "nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);");