+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - select user for protected region in team folder
 - edit xlsb
 - customization_goback setting removed
+- there are only system, light and dark themes
 
 ## 9.9.0
 ## Added